Add JavaByteSerialization based on SerializationTest
