+++ jvm/src/test/scala/scala/xml/SerializationTest.scala
@@ -1,38 +1,18 @@
 package scala.xml
 
-import java.io._
-
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
 class SerializationTest {
-  def roundTrip[T](obj: T): T = {
-    def serialize(in: T): Array[Byte] = {
-      val bos = new ByteArrayOutputStream()
-      val oos = new ObjectOutputStream(bos)
-      oos.writeObject(in)
-      oos.flush()
-      bos.toByteArray()
-    }
-
-    def deserialize(in: Array[Byte]): T = {
-      val bis = new ByteArrayInputStream(in)
-      val ois = new ObjectInputStream(bis)
-      ois.readObject.asInstanceOf[T]
-    }
-
-    deserialize(serialize(obj))
-  }
-
   @Test
   def xmlLiteral: Unit = {
     val n = <node/>
-    assertEquals(n, roundTrip(n))
+    assertEquals(n, JavaByteSerialization.roundTrip(n))
   }
 
   @Test
   def empty: Unit = {
-    assertEquals(NodeSeq.Empty, roundTrip(NodeSeq.Empty))
+    assertEquals(NodeSeq.Empty, JavaByteSerialization.roundTrip(NodeSeq.Empty))
   }
 
   @Test
@@ -40,6 +20,6 @@ class SerializationTest {
     val parent = <parent><child></child><child/></parent>
     val children: Seq[Node] = parent.child
     val asNodeSeq: NodeSeq = children
-    assertEquals(asNodeSeq, roundTrip(asNodeSeq))
+    assertEquals(asNodeSeq, JavaByteSerialization.roundTrip(asNodeSeq))
   }
 }
